updated runtime registration
