+++ .github/workflows/node_engine_artifacts.yml
@@ -24,7 +24,7 @@ jobs:
         run: |
           cd go_node_engine
           env GOOS=linux GOARCH=${{ matrix.architecture }}  go build -ldflags="-X 'go_node_engine/cmd.Version=${{ github.ref_name }}'" -o ../NodeEngine NodeEngine.go
-          env GOOS=linux GOARCH=${{ matrix.architecture }}  go build -o ../nodeengined pkg/nodeengined.go
+          env GOOS=linux GOARCH=${{ matrix.architecture }}  go build -o ../nodeengined internal/daemon/nodeengined.go
 
       - name: Package Artifacts
         run: |
